* question classification task finished

+++ models/TextCNN.py
@@ -26,7 +26,7 @@ class TextCNN(object):
 
         # Placeholders for input, output and dropout
         self.input_x = tf.placeholder(tf.int32, [None, sequence_length], name="input_x")
-        self.input_y = tf.placeholder(tf.float32, [None], name="input_y")
+        self.input_y = tf.placeholder(tf.int64, [None], name="input_y")
         self.dropout_keep_prob = tf.placeholder(tf.float32, name="dropout_keep_prob")
 
         # Keeping track of L2 regularization loss (optional)
